requirements.txt
fastapi
uvicorn
motor
# Non essential
python-dotenv
